Retro Task Tracker v0.1
Instructions for user:

1. Adding a task:
Enter the content of the new task in the "New task" field.
Click the "Add task."
Marking

2. completing the task:
Click the square next to a task to mark or deselect its completion.

3. Delete a task:
Click the "Delete" button next to a task to delete it.

4. Saving your task list:
Select "Save" on the "File" menu to save your current task list.

5. Loading the task list:
Select "Load" in the "File" menu to load a previously saved task list.

6. Change Theme:
Select your preferred theme from the "Theme" menu (Classic, Light, Dark).
------------------------------------------------------------------------

The Task Tracker app allows you to track and manage your task list. Tasks are numbered and you can add, delete and mark their completion. Progress in completing tasks is shown using a progress bar. It is also possible to customize the appearance of the application by changing the theme.

------------------------------------------------------------------------
Things I want to do in future updates:
>Language menu
>Saving task files in different file formats like pdf, docx, txt, etc.
>Better menu look
>More interactive UI